Misc. portability fixes, esp. plain char signedness.
P4:32624

+++ src/orl/test/c/test.c
@@ -414,8 +414,8 @@ orl_return PrintSymTable( orl_sec_handle orl_sec_hnd )
     return( ORL_OKAY );
 }
 
-static void * objRead( void *hdl, unsigned int len )
-/**************************************************/
+static void * objRead( void *hdl, size_t len )
+/********************************************/
 {
     buff_list   ptr;
